refactor: RecentBlogPosts widget #308
