+++ src/Core/Fan.WebApp/Fan.WebApp.csproj
@@ -27,6 +27,7 @@
     <ProjectReference Include="..\..\Widgets\BlogArchives\BlogArchives.csproj" />
     <ProjectReference Include="..\..\Widgets\BlogCategories\BlogCategories.csproj" />
     <ProjectReference Include="..\..\Widgets\BlogTags\BlogTags.csproj" />
+    <ProjectReference Include="..\..\Widgets\RecentBlogPosts\RecentBlogPosts.csproj" />
     <ProjectReference Include="..\Fan.Web\Fan.Web.csproj" />
   </ItemGroup>
 
@@ -35,6 +36,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Widgets\RecentBlogPosts\" />
     <Folder Include="wwwroot\plugins\shortcodes\" />
   </ItemGroup>
