+++ src/Themes/Clarity/scss/_widgets.scss
@@ -165,6 +165,10 @@ aside .recent-blog-posts {
     li {
         list-style: square;
     }
+
+    p {
+        font-size: smaller;
+    }
 }
 
 .blog-afterpost .recent-blog-posts {
