+++ src/Widgets/RecentBlogPosts/Components/RecentBlogPosts.cshtml
@@ -1,5 +1,5 @@
-﻿@using Fan.Blog.Models
-@using Fan.WebApp.Widgets.RecentBlogPosts
+﻿@using RecentBlogPosts.Components
+@using RecentBlogPosts
 @model Tuple<List<RecentPostViewModel>, RecentBlogPostsWidget>
 @{
     var posts = Model.Item1;
