+++ src/Widgets/RecentBlogPosts/Components/RecentPostViewModel.cs
@@ -1,0 +1,11 @@
+﻿namespace RecentBlogPosts.Components
+{
+    public class RecentPostViewModel
+    {
+        public string Title { get; set; }
+        public string Link { get; set; } 
+        public string Author { get; set; }
+        public string Date { get; set; }
+        public string Excerpt { get; set; }
+    }
+}
