+++ src/Widgets/RecentBlogPosts/Manage/Widgets/RecentBlogPostsEdit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model RecentBlogPostsEditModel
+@model RecentBlogPosts.Manage.Widgets.RecentBlogPostsEditModel
 @{
     ViewData["Title"] = "RecentBlogPosts Widget Settings";
     Layout = "_SettingsLayout";
